+++ scripts/avail-ext-doc/main.py
@@ -32,10 +32,10 @@ SCRIPTS_LOCATION = os.path.join(REPO_LOCATION, 'scripts')
 REPO_CLI_DOCS = 'https://{}@github.com/{}'.format(GH_TOKEN, DOC_REPO_SLUG)
 CLONED_CLI_DOCS = os.path.join(os.path.sep, 'doc-repo')
 AVAILABLE_EXTENSIONS_DOC = os.path.join(CLONED_CLI_DOCS, 'docs-ref-conceptual', 'azure-cli-extensions-list.md')
-TEMPLATE_FILE = "list-template.md"
+TEMPLATE_FILE = os.path.join(SCRIPTS_LOCATION, "avail-ext-doc", "list-template.md")
 
 sys.path.insert(0, SCRIPTS_LOCATION)
-from ci.util import get_index_data
+from ci.util import get_index_data, INDEX_PATH
 
 def get_extensions():
     extensions = []
@@ -52,26 +52,26 @@ def get_extensions():
         })
     return extensions
 
-def update_extensions_list(template_file, output_file):
+def update_extensions_list(output_file):
     template = None
-    with open(template_file, 'r') as doc_template:
+    with open(TEMPLATE_FILE, 'r') as doc_template:
         template = Template(doc_template.read())
     if template is None:
         raise RuntimeError("Failed to read template file {}".format(template_file))
     with open(output_file, 'w') as output:
         output.write(template.render(extensions=get_extensions(), date=datetime.date.today().strftime("%m/%d/%Y")))
 
-def date_only_diff(diff):
-    if not diff:
-        return True
-    stats = diff.split('\t')
-    return stats[0] == 1 and stats[1] == 1
+def needs_update(doc_repo):
+    date_format="%Y-%m-%d %H:%M:%S %z"
+    ext_repo = Repo(REPO_LOCATION)
+    doc_updated = datetime.datetime.strptime(doc_repo.git.log(AVAILABLE_EXTENSIONS_DOC, pretty="format:%ai", n=1), date_format)
+    template_updated = datetime.datetime.strptime(ext_repo.git.log(TEMPLATE_FILE, pretty="format:%ai", n=1), date_format)
+    index_updated = datetime.datetime.strptime(ext_repo.git.log(INDEX_PATH, pretty="format:%ai", n=1), date_format)
+
+    return (doc_updated < index_updated) or (doc_updated < template_updated)
 
 def commit_update(doc_repo):
     doc_repo.git.add(doc_repo.working_tree_dir)
-    if date_only_diff(doc_repo.git.diff(staged=True, numstat=True)):
-        logger.warning('No changes. Exiting')
-        return
     github_con = Github(GH_TOKEN)
     user = github_con.get_user()
     doc_repo.git.config('user.email', user.email or 'azpycli@microsoft.com')
@@ -94,8 +94,11 @@ def commit_update(doc_repo):
 
 def main():
     doc_repo = Repo.clone_from(REPO_CLI_DOCS, CLONED_CLI_DOCS)
-    update_extensions_list(TEMPLATE_FILE, AVAILABLE_EXTENSIONS_DOC)
-    commit_update(doc_repo)
+    if needs_update(doc_repo):
+        update_extensions_list(AVAILABLE_EXTENSIONS_DOC)
+        commit_update(doc_repo)
+    else:
+        logger.info("No update to docs index required, skipping publish")
 
 if __name__ == '__main__':
     main()