+++ src/find/azext_find/custom.py
@@ -6,31 +6,40 @@
 import random
 import json
 import re
+import sys
+import platform
 import requests
+import colorama  # pylint: disable=import-error
 
 
 from azure.cli.core import telemetry as telemetry_core
-
+from knack.prompting import prompt
 from knack.log import get_logger
 logger = get_logger(__name__)
 
-wait_messages = ['Ok, let me find an answer to that question.',
-                 'I\'m working on finding the right answer for you.', 'Let me see if I can answer that for you.']
+WAIT_MESSAGE = ['I\'m an AI bot (learn more: aka.ms/aladdin); Let me see how I can help you...']
+
+EXTENSION_NAME = 'find'
+
+FIND_EXTENSION_PREFIX = 'Context.Default.Extension.Find.'
 
 
-def process_query(question):
-    print(random.choice(wait_messages))
-    response = call_aladdin_service(question)
+def process_query(cli_term):
+    print(random.choice(WAIT_MESSAGE))
+    response = call_aladdin_service(cli_term)
 
     if response.status_code != 200:
         logger.error('[?] Unexpected Error: [HTTP {0}]: Content: {1}'.format(response.status_code, response.content))
     else:
+        if (platform.system() == 'Windows' and should_enable_styling()):
+            colorama.init(convert=True)
+
         answer_list = json.loads(response.content)
         if (not answer_list or answer_list[0]['source'] == 'bing'):
-            logger.warning("Sorry I am not able to help with that. \nTry typing the beginning of a "
-                           "command e.g. 'az vm' or explain the task you want to accomplish e.g. 'create a vm'.")
+            print("\nSorry I am not able to help with [" + cli_term + "]."
+                  "\nTry typing the beginning of a command e.g. " + style_message('az vm') + ".")
         else:
-            print("Here are some information I was able to gather for you: \n")
+            print("\nHere are the most common ways to use [" + cli_term + "]: \n")
             num_results_to_show = min(3, len(answer_list))
             for i in range(num_results_to_show):
                 current_title = answer_list[i]['title'].strip()
@@ -43,11 +52,49 @@ def process_query(question):
                     current_snippet = current_snippet[start_index:]
                 current_snippet = current_snippet.replace('```', '').replace(current_title, '').strip()
                 current_snippet = re.sub(r'\[.*\]', '', current_snippet).strip()
-                print('\033[1m' + current_title + '\033[0m')
+                print(style_message(current_title))
                 print(current_snippet)
 
-                if i + 1 < num_results_to_show:
-                    print("\n")
+                print("")
+            feedback = prompt("[Enter to close. Press + or - to give feedback]:")
+            if feedback in ['+', '-']:
+                print('Wow, you are a true hero!')
+                print("""\
+        O_
+       """ + style_message("""<T>""") + """`-.
+        """ + style_message("""|""") + """`-‘
+        """ + style_message("""I""") + """
+                        """)
+                print('My human overlords review each of these reports; I\'m told these reports makes me smarter.')
+                print('Send us more feedback by email: aladdindoc@microsoft.com')
+            properties = {}
+            set_custom_properties(properties, 'Feedback', feedback)
+            telemetry_core.add_extension_event(EXTENSION_NAME, properties)
+
+
+def style_message(msg):
+    if should_enable_styling():
+        try:
+            msg = colorama.Style.BRIGHT + msg + colorama.Style.RESET_ALL
+        except KeyError:
+            pass
+    return msg
+
+
+def should_enable_styling():
+    try:
+        # Style if tty stream available
+        if sys.stdout.isatty():
+            return True
+    except AttributeError:
+        pass
+    return False
+
+
+def set_custom_properties(prop, name, value):
+    if name and value is not None:
+        # 10 characters limit for strings
+        prop['{}{}'.format(FIND_EXTENSION_PREFIX, name)] = value[:10] if isinstance(value, str) else value
 
 
 def call_aladdin_service(query):
@@ -58,9 +105,6 @@ def call_aladdin_service(query):
         'installation_id': telemetry_core._get_installation_id()  # pylint: disable=protected-access
     }
 
-    if (query and query.startswith("az ")):
-        query = '"' + query + '"'
-
     service_input = {
         'paragraphText': "<div id='dummyHeader'></div>",
         'currentPageUrl': "",
@@ -68,7 +112,7 @@ def call_aladdin_service(query):
         'context': context
     }
 
-    api_url = 'https://aladdinservice-staging.azurewebsites.net/api/aladdin/generateCards'
+    api_url = 'https://aladdinservice-prod.azurewebsites.net/api/aladdin/generateCards'
     headers = {'Content-Type': 'application/json'}
 
     response = requests.post(api_url, headers=headers, json=service_input)