+++ src/image-copy/azext_imagecopy/__init__.py
@@ -43,6 +43,12 @@ class ImageCopyCommandsLoader(AzCommandsLoader):
                        help='Name of the final image that will be created')
             c.argument('target_subscription', options_list=['--target-subscription'],
                        help='Name or ID of the subscription where the final image should be created.')
+            c.argument('timeout', options_list=['--timeout'], type=int, default=3600,
+                       help='Time in seconds for the copy operation to finish. Increase this time if '
+                       'you are going to copy large images (disks) like 512GB or more.')
+            c.argument('temporary_resource_group_name', options_list=['--temporary_resource_group_name'],
+                       default='image-copy-rg',
+                       help='Resource Group name where temporary storage account will be created.')
 
 
 COMMAND_LOADER_CLS = ImageCopyCommandsLoader