+++ src/mesh/azext_mesh/servicefabricmesh/mgmt/servicefabricmesh/models/service_resource_description.py
@@ -26,10 +26,12 @@ class ServiceResourceDescription(ManagedProxyResource):
     :ivar type: The type of the resource. Ex-
      Microsoft.Compute/virtualMachines or Microsoft.Storage/storageAccounts.
     :vartype type: str
-    :param os_type: The Operating system type required by the code in service.
-     . Possible values include: 'Linux', 'Windows'
+    :ivar provisioning_state: State of the resource.
+    :vartype provisioning_state: str
+    :param os_type: The operation system required by the code in service.
+     Possible values include: 'Linux', 'Windows'
     :type os_type: str or
-     ~azure.mgmt.servicefabricmesh.models.OperatingSystemTypes
+     ~azure.mgmt.servicefabricmesh.models.OperatingSystemType
     :param code_packages: Describes the set of code packages that forms the
      service. A code package describes the container and the properties for
      running it. All the code packages are started together on the same host
@@ -46,47 +48,67 @@ class ServiceResourceDescription(ManagedProxyResource):
     :param replica_count: The number of replicas of the service to create.
      Defaults to 1 if not specified.
     :type replica_count: int
-    :param health_state: The health state of a resource such as Application,
-     Service, or Network. Possible values include: 'Invalid', 'Ok', 'Warning',
-     'Error', 'Unknown'
-    :type health_state: str or
-     ~azure.mgmt.servicefabricmesh.models.HealthState
-    :ivar status: Represents the status of the service. Possible values
-     include: 'Unknown', 'Active', 'Upgrading', 'Deleting', 'Creating',
-     'Failed'
+    :param auto_scaling_policies: Auto scaling policies
+    :type auto_scaling_policies:
+     list[~azure.mgmt.servicefabricmesh.models.AutoScalingPolicy]
+    :ivar status: Status of the service. Possible values include: 'Unknown',
+     'Ready', 'Upgrading', 'Creating', 'Deleting', 'Failed'
     :vartype status: str or
-     ~azure.mgmt.servicefabricmesh.models.ServiceResourceStatus
+     ~azure.mgmt.servicefabricmesh.models.ResourceStatus
+    :ivar status_details: Gives additional information about the current
+     status of the service.
+    :vartype status_details: str
+    :ivar health_state: Describes the health state of an application resource.
+     Possible values include: 'Invalid', 'Ok', 'Warning', 'Error', 'Unknown'
+    :vartype health_state: str or
+     ~azure.mgmt.servicefabricmesh.models.HealthState
+    :ivar unhealthy_evaluation: When the service's health state is not 'Ok',
+     this additional details from service fabric Health Manager for the user to
+     know why the service is marked unhealthy.
+    :vartype unhealthy_evaluation: str
     """
 
     _validation = {
         'id': {'readonly': True},
         'type': {'readonly': True},
+        'provisioning_state': {'readonly': True},
         'os_type': {'required': True},
         'code_packages': {'required': True},
         'status': {'readonly': True},
+        'status_details': {'readonly': True},
+        'health_state': {'readonly': True},
+        'unhealthy_evaluation': {'readonly': True},
     }
 
     _attribute_map = {
         'id': {'key': 'id', 'type': 'str'},
         'name': {'key': 'name', 'type': 'str'},
         'type': {'key': 'type', 'type': 'str'},
+        'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'os_type': {'key': 'properties.osType', 'type': 'str'},
         'code_packages': {'key': 'properties.codePackages', 'type': '[ContainerCodePackageProperties]'},
         'network_refs': {'key': 'properties.networkRefs', 'type': '[NetworkRef]'},
         'diagnostics': {'key': 'properties.diagnostics', 'type': 'DiagnosticsRef'},
         'description': {'key': 'properties.description', 'type': 'str'},
         'replica_count': {'key': 'properties.replicaCount', 'type': 'int'},
-        'health_state': {'key': 'properties.healthState', 'type': 'str'},
+        'auto_scaling_policies': {'key': 'properties.autoScalingPolicies', 'type': '[AutoScalingPolicy]'},
         'status': {'key': 'properties.status', 'type': 'str'},
+        'status_details': {'key': 'properties.statusDetails', 'type': 'str'},
+        'health_state': {'key': 'properties.healthState', 'type': 'str'},
+        'unhealthy_evaluation': {'key': 'properties.unhealthyEvaluation', 'type': 'str'},
     }
 
-    def __init__(self, os_type, code_packages, name=None, network_refs=None, diagnostics=None, description=None, replica_count=None, health_state=None):
+    def __init__(self, os_type, code_packages, name=None, network_refs=None, diagnostics=None, description=None, replica_count=None, auto_scaling_policies=None):
         super(ServiceResourceDescription, self).__init__(name=name)
+        self.provisioning_state = None
         self.os_type = os_type
         self.code_packages = code_packages
         self.network_refs = network_refs
         self.diagnostics = diagnostics
         self.description = description
         self.replica_count = replica_count
-        self.health_state = health_state
+        self.auto_scaling_policies = auto_scaling_policies
         self.status = None
+        self.status_details = None
+        self.health_state = None
+        self.unhealthy_evaluation = None