+++ src/mesh/azext_mesh/servicefabricmesh/mgmt/servicefabricmesh/operations/network_operations.py
@@ -22,7 +22,7 @@ class NetworkOperations(object):
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
     :param deserializer: An object model deserializer.
-    :ivar api_version: The version of the API. This parameter is required and its value must be `2018-07-01-preview`. Constant value: "2018-07-01-preview".
+    :ivar api_version: The version of the API. This parameter is required and its value must be `2018-09-01-preview`. Constant value: "2018-09-01-preview".
     """
 
     models = models
@@ -32,27 +32,24 @@ class NetworkOperations(object):
         self._client = client
         self._serialize = serializer
         self._deserialize = deserializer
-        self.api_version = "2018-07-01-preview"
+        self.api_version = "2018-09-01-preview"
 
         self.config = config
 
     def create(
-            self, resource_group_name, network_name, network_resource_description, custom_headers=None, raw=False, **operation_config):
+            self, resource_group_name, network_resource_name, network_resource_description, custom_headers=None, raw=False, **operation_config):
         """Creates or updates a network resource.
 
-        Creates a network resource with the specified name and description. If
-        a network with the same name already exists, then its description is
-        updated to the one indicated in this request.
-        Use network resources to create private network and configure public
-        connectivity for services within your application.
-        .
+        Creates a network resource with the specified name, description and
+        properties. If a network resource with the same name exists, then it is
+        updated with the specified description and properties.
 
         :param resource_group_name: Azure resource group name
         :type resource_group_name: str
-        :param network_name: The identity of the network.
-        :type network_name: str
+        :param network_resource_name: The identity of the network.
+        :type network_resource_name: str
         :param network_resource_description: Description for creating a
-         network resource.
+         Network resource.
         :type network_resource_description:
          ~azure.mgmt.servicefabricmesh.models.NetworkResourceDescription
         :param dict custom_headers: headers that will be added to the request
@@ -72,7 +69,7 @@ class NetworkOperations(object):
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self.config.subscription_id", self.config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'networkName': self._serialize.url("network_name", network_name, 'str', skip_quote=True)
+            'networkResourceName': self._serialize.url("network_resource_name", network_resource_name, 'str', skip_quote=True)
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -98,7 +95,7 @@ class NetworkOperations(object):
         response = self._client.send(
             request, header_parameters, body_content, stream=False, **operation_config)
 
-        if response.status_code not in [200, 201]:
+        if response.status_code not in [200, 201, 202]:
             raise models.ErrorModelException(self._deserialize, response)
 
         deserialized = None
@@ -113,21 +110,20 @@ class NetworkOperations(object):
             return client_raw_response
 
         return deserialized
-    create.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkName}'}
+    create.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkResourceName}'}
 
     def get(
-            self, resource_group_name, network_name, custom_headers=None, raw=False, **operation_config):
-        """Gets the network resource.
+            self, resource_group_name, network_resource_name, custom_headers=None, raw=False, **operation_config):
+        """Gets the network resource with the given name.
 
-        Gets the information about the network resource with a given name. This
-        information includes the network description and other runtime
-        information.
-        .
+        Gets the information about the network resource with the given name.
+        The information include the description and other properties of the
+        network.
 
         :param resource_group_name: Azure resource group name
         :type resource_group_name: str
-        :param network_name: The identity of the network.
-        :type network_name: str
+        :param network_resource_name: The identity of the network.
+        :type network_resource_name: str
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -145,7 +141,7 @@ class NetworkOperations(object):
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self.config.subscription_id", self.config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'networkName': self._serialize.url("network_name", network_name, 'str', skip_quote=True)
+            'networkResourceName': self._serialize.url("network_resource_name", network_resource_name, 'str', skip_quote=True)
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -180,18 +176,18 @@ class NetworkOperations(object):
             return client_raw_response
 
         return deserialized
-    get.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkName}'}
+    get.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkResourceName}'}
 
     def delete(
-            self, resource_group_name, network_name, custom_headers=None, raw=False, **operation_config):
+            self, resource_group_name, network_resource_name, custom_headers=None, raw=False, **operation_config):
         """Deletes the network resource.
 
         Deletes the network resource identified by the name.
 
         :param resource_group_name: Azure resource group name
         :type resource_group_name: str
-        :param network_name: The identity of the network.
-        :type network_name: str
+        :param network_resource_name: The identity of the network.
+        :type network_resource_name: str
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the
          deserialized response
@@ -207,7 +203,7 @@ class NetworkOperations(object):
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self.config.subscription_id", self.config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str'),
-            'networkName': self._serialize.url("network_name", network_name, 'str', skip_quote=True)
+            'networkResourceName': self._serialize.url("network_resource_name", network_resource_name, 'str', skip_quote=True)
         }
         url = self._client.format_url(url, **path_format_arguments)
 
@@ -229,22 +225,21 @@ class NetworkOperations(object):
         request = self._client.delete(url, query_parameters)
         response = self._client.send(request, header_parameters, stream=False, **operation_config)
 
-        if response.status_code not in [200, 204]:
+        if response.status_code not in [200, 202, 204]:
             raise models.ErrorModelException(self._deserialize, response)
 
         if raw:
             client_raw_response = ClientRawResponse(None, response)
             return client_raw_response
-    delete.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkName}'}
+    delete.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ServiceFabricMesh/networks/{networkResourceName}'}
 
     def list_by_resource_group(
             self, resource_group_name, custom_headers=None, raw=False, **operation_config):
         """Gets all the network resources in a given resource group.
 
         Gets the information about all network resources in a given resource
-        group. The information includes the network description and other
-        runtime properties.
-        .
+        group. The information include the description and other properties of
+        the Network.
 
         :param resource_group_name: Azure resource group name
         :type resource_group_name: str
@@ -313,9 +308,9 @@ class NetworkOperations(object):
             self, custom_headers=None, raw=False, **operation_config):
         """Gets all the network resources in a given subscription.
 
-        Gets the information about all network resources in a given
-        subscription. The information includes the network description and
-        other runtime properties.
+        Gets the information about all network resources in a given resource
+        group. The information include the description and other properties of
+        the network.
 
         :param dict custom_headers: headers that will be added to the request
         :param bool raw: returns the direct response alongside the