+++ src/webapp/azext_webapp/_constants.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-NODE_VERSION_DEFAULT = "8.9"
+NODE_VERSION_DEFAULT = "8.11"
 NETCORE_VERSION_DEFAULT = "2.0"
 DOTNET_VERSION_DEFAULT = "4.7"
 PYTHON_VERSION_DEFAULT = "3.7"
@@ -14,6 +14,6 @@ PYTHON_RUNTIME_NAME = "python"
 OS_DEFAULT = "Windows"
 STATIC_RUNTIME_NAME = "static"  # not an oficial supported runtime but used for CLI logic
 # TODO: Remove this once we have the api returning the versions
-NODE_VERSIONS = ['4.4', '4.5', '6.2', '6.6', '6.9', '6.11', '8.0', '8.1']
+NODE_VERSIONS = ['4.4', '4.5', '6.2', '6.6', '6.9', '6.11', '8.0', '8.1', '8.9', '8.11']
 NETCORE_VERSIONS = ['1.0', '1.1', '2.0']
 DOTNET_VERSIONS = ['3.5', '4.7']