+++ src/mesh/HISTORY.rst
@@ -3,6 +3,20 @@
 Release History
 ===============
 
+0.10.2 (2018-11-27)
+++++++++++++++++++
+
+* Added yaml support
+* simplified deployment logic
+
+0.9.4 (2018-10-19)
+++++++++++++++++++
+
+* Added gateway/network changes.
+* Changed api version to 2018-09-01-preview.
+* Fixed issue with displaying error messages
+
+
 0.9.3 (2018-08-03)
 ++++++++++++++++++